\name{format.relation}
\alias{format.relation}
\title{
  Internal utility to format the matrix used to specify special 
  relationships between pairs of individuals
}
\description{
  Internal utility to format the matrix used to specify special 
  relationships between pairs of individuals. 
  See kinship2::pedigree for details. 
}
\usage{
format.relation(relation)
}
\arguments{
  \item{relation}{A matrix for one or more special relationships}
}
\value{
  A matrix for one or more special relationships as required by pedgene  
}
\author{
Alessia Visconti, King's College London (alessia.visconti@kcl.ac.uk).
}
\seealso{
pedgene, kinship
}
